Alice does stage 1, sends to Bob.  Bob does stage 2, sends back to Alice.
The sms part is surely broken (not like I can test it anyway)
There is a lot of refactoring to do.  I want to find what Alice and Bob functions for stage 1, 2, 3, and 4 have it common and write functions in protocol that do that stuff.  End result being aliceStage1, aliceStage3 etc but they won't repeat code, they'll use functions with different input.